Use "this" keyword to get Marker object
Using the index of the target to get the index of the Marker object we need from the markers array was not working right because I added additional markers to my map in order to implement a current location marker feature. These additional markers changed the amount of marker elements on the map. Because there are now more markers on the map than there are in the array that's passed to setCollection, we can no longer rely on the indexes of each list being the same. Passing my new Markers to setCollection wouldn't work well because it shouldn't be included with clustering or spidering like the others are. Luckily we don't need to do it this. The Google Maps API binds the "this" keyword in the event listener callback to the Marker object that we want to pass, so all we have to do is pass "this" and we've passed the Marker we need.